refactor(single-instance): move singleton.py → os_/single_instance.py
